Save proclist that can be used as example model proclist.
In order to test the function `format_model_data`, I saved a paired down version of an actual proclist produced by `get_clusterassignment`. The paired down  only includes the variables that are used by `format_model_data`. The values for these variables are the same as in the actual proclist.