Updated `analyze_questioned_documents` to output the likelihoods, votes, and posterior proabilities.
The posterior probabilities are most likely what the end user will want, but I'm including the votes and likelihoods in case the user is also interested in those.

I saved the output of  `analyze_questioned_documents` on the `example_model_training_data` and `example_questioned_data` as `example_analysis` in the data folder so that it can be used in examples and testing.

I wrote very basic tests for `analyze_questioned_documents`. Right now the test check that the output is a named list with names "likelihoods", "votes", and "posterior_probabilities". Tests also check that the number of votes for each document sum to the number of MCMC iterations (with burn-in dropped) and that the posterior probabilities for each document sum to 1.