Make `make_clustering_templates()` easier to debug.
The function `make_clustering_templates()` initially consisted of nested functions four+ levels deep. This was how Amy structured the code originally. My original goal working with Amy's code was to only make strutural changes necessary to get the code running on new input data, so I hadn't changed the nesting. Now that the code is working on new input data, I decided to remove all but one layer of nesting. This makes it easier to make enhancements to the code.

The names of the training documents used to create template(s) are now stored in the output template(s).