Fixed bug. Writer IDs caused error in model.
I had changed docwriter and letterwriter to be the writer IDs, but this caused the following error in the model

"Error in rjags::jags.model(textConnection(model_wrapped_cauchy), data = model_training_data,  :
  RUNTIME ERROR:
Compilation error on line 25.
Index out of range taking subset of  theta"

Changing docwriter and letterwriter to number the writers 1,2,...,W fixed the error.

I updated the tests to check that docwriter and letterwriter are contain integers 1,2,...,W.