Fixed bug. Error in model when clusters are numbered sequentially.
I found an example where the template has 10 clusters but the model training data and questioned data only fill clusters 1,2,4,5,6,7,8,10. Because the clusters aren't labeled sequentially the model throughs an index out of bounds error because it is trying to use the cluster labels as indices.

I changed format_model_data() and format_questioned_data() to relabel the clusters sequentially if necessary. The model training data is relabled first. format_questioned_data() detects if the model data clusters were relabled and relables the questioned data clusters accordingly.

An error will still occur if the questioned data uses a cluster that the model training data doesn't use. I will open a separate issue for that because I need to check with Danica and Alicia to see how we should handle that case.