Update documentation of example data.
I updated the documentation of the example data after creating new example data with a template with K=10 clusters instead of K=5.

I changed the number of clusters because I thought that for K=5 the model docs didn't assign any graphs to cluster #5 but the questioned docs assigned quite a few to this cluster. But I later discovered that I had been using the wrong questioned docs. However, I had already recreated the example data with K=10 and everything worked, so there's no reason to switch back to K=5.