Change number of cores to 2 to pass check().
I initially had the number of cores in `analyze_questioned_documents` set to 4 in both the vignette and the test. This caused an error

`4 simultaneous processes spawned`

when running R CMD build. Stackoverflow https://stackoverflow.com/questions/50571325/r-cran-check-fail-when-using-parallel-functions says that CRAN limits the number of cores to 2, so I updated the vignette and test accordingly.