Update cluster template to store writer for each graph.
In order to measure how well the template distinguishes between training writers, we need to know which graph came from which writer. I updated the make_cluster_templates() function to store a vector of writers that is the same length as the vector of clusters.

In the future it might make more sense to store graph #, writer ID, and cluster assignment in a data frame instead of separate vectors. I didn't do that here because doing so would require that other funcitons in the pipeline that use the clusters vector be updated.