Add traceplots and more chains.
## Traceplots
I changed the output of fit_model() to be the output of coda.samples() without formatting the draws so that the user could run coda::traceplot() on the output. coda::traceplot() creates the traceplot for every single variable. I wrote a function plot_trace() to create a traceplot for a single variable.

I wrote a helper funciton format_draws() to take the output of fit_model() and format for the next step in the model pipeline.

## Chains
I updated fit_model() to allow the user to run more than one MCMC chain.

## Tests
Tests are currently failing. I haven't updated them yet.