Updated examples.
After changing the output format of fit_model() and drop_burnin(). I updated the examples. In an attempt to make things easier, I started using the ROXygen tag `@inherit fit_model examples` to reuse examples insted of copying and pasting examples. My hope is that now if examples need to be updated, there are fewer places where changes need to be made.