## New button I added a button to process template images in the shiny app. The processed images are automatically saved to q_main_dir/data/template_graphs.
## Changed process_batch_dir()
I changed process_batch_dir() so that as soon as an image is processed it is saved instead of waiting until all images are processed and saving the images at the end. This way no progress is lost if one of the images can't be processed.

I also changed the function to check with the image has already be processed and saved to the output directory. If so, this image is skipped. After all input images have been processed or skipped, the function loads all processed images in the output directory into a list and returns the list. This way the function won't process the same image multiple times (in the same output dir anyways).