Turned off extra tabs in Shiny App
extractFeatures doesn't isn't set up yet to work with the next step in the analysis pipeline so I turned off this tab.

clustering is now done on the Analyze Questioned Documents tab so I turned it off.

The triangleDecomposition and slr tabs are just place holders so I turned them off.