Add option to use a default template in Shiny app.
Alicia asked that I add the option to use a default template. For right now at least, I am using example_cluster_template that is included in the handwriter package as the default template because it only has K=10 clusters so extracting cluster assignments for the model and questioned data will go quickly. In the future, the default template should be much more robust.