Clean up. Changed button names to get and save model data in Shiny app.
I changed the button "q_get_model_clusters" to get "q_get_model_data" because the end result of the button is the output of format_model_data(). However, I left the button label as "Get Cluster Assignments" because this is what the user is likely interested in and "Get Data" is too vague.