Add cluster fill count plots to Shiny app.
## Current
I added code to generate cluster fill count plots in the analyze.R server script. These plots are displayed in the model and questioned tabs, and are generated when the get cluster assigments buttons are pushed.

## Future work
A function should be created and added to handwriter to generate cluster fill count plots. After that function is in handwriter. The shiny server script analyze.R should be updated to use that function.