Relabeled some RJAGS model parameters to match AOAS paper notation.
## Relabeled
I relabeled

- nll_datamodel as theta
- rho as tau
- theta as pi
- pi as pi_constant

so that the RJAGS model parameters are the same as the AOAS paper. This should make it easier for users to understand the model diagnostics.

## Did not relabel
I think nld_locationparam should be relabeled as mu, but I'm not certain. I don't understand why mu instead of nld_locationparam is used in the wrapped-cauchy distribution for theta. I won't make any changes to nld_locationparam or mu until I hear back from Alicia or Danica.