Add tests for about_variable()
Wrote several basic tests for about_variable(). Also changed the function to return an error if the input variable isn't in the model.

I tried writing a test for the error:
expect_error(about_variable(variable = "eta[9]", model_data = example_model_data, model = model),
  'Error in about_variable(variable = "eta[9]", model_data = example_model_data,  :
  The input variable eta[9] is not in the model.')

but this returns

"Error in grepl(pattern, conditionMessage(x), ...) :
  invalid regular expression 'Error in about_variable(variable = "eta[9]", model_data = example_model_data,  :
  The input variable eta[9] is not in the model.', reason 'Missing ')''
In addition: Warning message:
In grepl(pattern, conditionMessage(x), ...) :
  TRE pattern compilation error 'Missing ')''"

I need to move on to other things, so I am omitting a test for the error for the time being.