Deleted rjags model text from ClusterModeling_modelfunctions.R
The rjags model text is flagged during check() because it isn't documented. The model is stored in an rda file in the data folder, so it isn't necessary to also define it in ClusterModeling_modelfunctions.R. I deleted the model from ClusterModeling_modelfunctions.R. and created a function in data-raw > example_data.R to save the model as an rda file in the data folder in case the rda file needs to be changed or recreated.