Write vignette draft explaining handwriter's implementation of K-means.
While looking at make_clustering_tempaltes(), I found myself trying to remember why I wrote the K-means algorithm to check the stopping conditions after the assignment step and before the update step. I did some digging through my notes, and recalled the reason. This seemed like a good time to write a vignette that explains this and some of the other choices we made in the implementation of K-means.