Add seeds to chooseGraphs() and chooseCenters().
Initially, I set the seed for the random number generator inside chooseCenters(). This had a bug: if I created two new cluster templates with the same training docs and seed and 1000 graphs, the starting cluster centers are identical, but the resulting templates are not identical. After some digging, I figured out that this is because the 1000 graphs are not identical. I created a new function, chooseGraphs(), to select a sample of graphs. I changed chooseCenters() and chooseGraphs() to use seeds.