Recreated example data with centers and graphs seeds.
Because make_clustering_templates() was updated to use two seeds - one for selecting cluster centers and one for selecting training graphs - instead of one, the example data needed to be recreated.

I also realized that the template folder in tests > testthat > fixtures needs to store the associated files for the new example data. When I initially set up the fixtures, I moved template files there by hand. I changed the example_data.R script to create the new example data in the template folder in the fixtures folder. This way the fixtures folder will always have the correct files for the template and model and no one will need to move files there by hand. However, not all of the files in the template folder are needed by the tests, so the last few lines in example_data.R delete the unnecessary files.

Created a function to calculate a model's accuracy on a test set of documents.

All the tests pass on the new example data. devtools::check(vignettes = FALSE) runs without errors, but does produce warnings and notes.