Created function to make samples from CSAFE docs
Created function select_csafe_documents() that builds three sets of documents:

- template training docs: to be used to create a new cluster tmeplate
- model training docs: to be used to fit a new model. The writers in this set are distinct from the template training writers.
- questioned (test) docs: to be used to test a new model. The writers are the same for this as as the model training writers but the documents are distinct between the two sets.

The function allows users to choose how many template training writers and model training writers to randomly select. The function also allows users to choose which prompts, sessions, and repetitions to use for each of the three sets. The function won't allow any of the same documents to be used in the model and the questioned sets.

I created a data object called csafe_docs which is a data frame of all docs in CSAFE Handwriting Database V.2. (This needs to be updated to the latest version.) The columns in the data frame are doc (file name), writer, session, prompt, and repetition. select_csafe_docs uses csafe_docs to get the info about available CSAFE docs.

I initially thought that I would write select_csafe_docs to copy the selected template, model, and questioned docs to output folders. But I think this is a little messy. Instead, I decided to have select_csafe_docs simply form the samples and return the doc names of the files in each sample. Next I will write a separate helper function to copy the desired docs to a new location.

I started a vignette to show how to use select_csafe_docs but haven't finished it yet.