Make a second analyze document function for speed comparison.
The problematic line is

`dwc_sums[, i] <- rowSums(t(sapply(1:niter, function(it) log(circular::dwrappedcauchy(x = circular::circular(m_pcrot), mu = circular::circular(layered_wc_params[it, m_cluster, 1]), rho = layered_wc_params[it, m_cluster, 2])))))
`

This line returns a warning that rho needs to be between 0 and 1 but the length is greater than 1. In other words, it is unhappy that rho is a vector of values between 0 and 1.

I'm afraid that rewriting the code with for loops so that circular::dwrappedcauchy is only called on scalars, not vectors, might slow down the code a lot. I created a new function analyze_questioned_documents2() that replaces the problematic line with scalar calls to circular::dwrappedcauchy so that I can compare the speed with analyze_questioned_documents(). Although, its best to keep in mind that even if analyze_questioned_documents() is faster on my computer, it doesn't actually run on Pronto.