Rewrite batch_process_dir() without purrr package
I thought the purrr package was causing an error so I tried rewriting the code to load a list of files using lapply instead of purrr. It turns out that one of the files I was trying to load was the problem, not purrr. However, because we want to try to reduce the number of packages that handwriter uses, I'm leaving the lapply statement in.