Changed get_clusterassignment() to save small dataframe instead of all graph info
The cluster assignment files end up being large because they store all of the original graph info about edges and nodes in addition to the cluster assignment. The only fields that are needed for the next step in the pipeline are the following:

- writer
- doc
- cluster
- slope
- xvar
- yvar
- covar

I changed the function to only output a data frame of these variables for each document.

I also changed the function to save a data frame of the above variables for each document in template_dir > data > model_clusters or questioned_clusters. The function now checks whether a data frame for that document already exists, and if so, it loads it instead of recalculating the cluster assignments. It can take a while to get cluster assignments for a large folder, so I think this might help in case the process gets stopped and restarted for some reason.