Fix the outfile name which was missing the file extension.
I thought docname contained the file extension (E.g., w0001_s01_pLND_r01.png) so I initially used str_replace to replace the .png with .rds. But docname does not contain the file extension so the file was being saved as w0001_s01_pLND_r01. I fixed it so that .rds is now added to docname for the outfile.