Create function to get and sort post probs for a single questioned document.
`analyze_questioned_documents` calculates the posterior probabilities for questioned document(s) and returns the results in a data frame sorted by known writer names. The new function `get_posterior_probabilities` selects the known writer column and the column for a particular questioned document from the larger posterior probabilities data frame and sorts the result from largest post prob to smallest. I.e. the known writer with the largest post prob for the questioned document will be in the first row.