Created function to count the number of correct top writers.
If the user fits a model using CSAFE writers and test (questioned) documents from those same CSAFE writers, the function `count_csafe_correct_top_writer` will count the number of questioned documents for which the highest posterior probability is correctly assigned to the true writer.

This function assumes that the test (questioned) documents are named with the format w1_s01_pLND_r01. The function extracts the writer ID from the document name by finding the number that follows the letter w. If a document does not use this naming convention, the function will fail.