Keep original node order and sorted nodes just in case
The function find_colorpoints() assigns colors to the nodes. In a previous commit, we fixed a bug with the blue and orange nodes by sorting words$terminalNodes. In that commit, I replaced words$terminalNodes with the sorted version. Just in case the original order of the nodes might be needed in the future, I changed the function to store the sorted nodes as a new variable.