Added functionality to sample questioned document graphs.
For the different lengths project, we want to be able to generate questioned documents with various numbers of graphs. I updated get_clusterassignment() to allow the user to choose the number of graphs to use from each input document. I also updated fit_model() to use only allow the user to use "All" graphs in get_clusterassignment() and I updated analyze_questioned_documents() and analyze_questioned_documents2() to allow the user to choose the number of graph they want to use from each questioned document.

Initially I had planned to not change any of the handwriter functions and instead generate qd of various lenghts by changing the graph rds files. But once I started looked at the code, it was going to be quite complicated to change the graph rds files and have them still work with the get_clusterassignment function. Changing the get_clusterassigment function is a much easier solution.