Created a new function to plot cluster fill rates added plot to Shiny app.
The model estimates the true cluster fill rate for each cluster and each model writer, these estimates are called the writer profiles. I added the plot of the writer profiles to the Shiny app. Because these plots show cluster fill rates instead of cluster fill counts, I created a new function to plot the raw cluster fill rates from the model training documents or the questioned documents so that these plots can be compared with the model writer profiles.