Display selected question document on the analyze tab.
Now the user may select a questioned document to display on the analyze_qd page of the Analyze tab. A drop-down menu will be updated with the names of the questioned docs in the qd folder and the user may choose which doc to display.By default, the first questioned document in the qd folder will be displayed when the user selects the qd folder.