Make numstrat an argument for make_clustering_templates()
The vector numstrat sets the number of graphs to sample from each stratum. Amy optimize numstrat for K=40. I need to figure out which numstrat works best for K=5, 10, 45, 50. To make this easier I changed make_clustering_templates() to take numstrat as a argument. As of right now, after I identify the 'optimal' numstrat for these K's, I plan to code the optimal numstrats into make_clustering_templates() drop it as an argument.