Fixed warning about missing link for getAllPairsDistances in man/getGraphDistance.Rd
getAllPairsDistances is not an exported function so it doesn't have a man file to link to. I removed the link in the ROxygen documentation and ran devtools::document() to update man/getGraphDistance.Rd.

devtools::check() produces 1 warning and 3 notes. The warning is, "checking whether package ‘handwriter’ can be installed ... [13s/13s] WARNING
  See below..."