Turned off Rdata and rda options in readPNGBinary because they don't work as is.
check() noted that "magick_image" does not have a global binding. When I went to the function to fix this note, I noticed that "magick_image" isn't defined anywhere in the function. I don't know an easy fix for this section of code. As far as I know, no one is using readPNGBinary to read images from Rdata or rda files anyway so I commented-out this section of code.

Fix the functionality that allows users to run readPNGBinary on Rdata or rda files. I think this is what enables the masking feature.