Changed `getGraphDistance()` to internal function and updated documention for functions in ClusterDistanceFunctions.R
## getGraphDistance
`getGraphDistance()` had been an exported function, but it requires the input graphs to be in a special format that would be difficult for users to replicate. In fact, without digging through the functions that call `getGraphDistance()` I would have no idea how to replicate the graph format. Thus I changed `getGraphDistance()` to an internal function instead of an exported one.

## Functions in ClusterDistanceFunctions.R
These functions are all internal so the only update I made to their documentation was to add the @noRd tag so that Rd files aren't generated for them.