Created `plotClusterTemplate()`
Created a new function to plot cluster templates. The function creates a plot for each cluster that displays the cluster's graphs and exemplars. Ideally, each cluster plot would be displayed as a subplot in a master plot. I couldn't figure out how to do this with ggplot, so the function returns a list of a cluster plots.

The patchwork package allows users to place ggplot plots as subplots in a larger plot. I didn't want to add another required package to handwriter, but I might add patchwork later as a suggested package.