Created new exported function to get cluster assignments
The internal function `get_clusterassignment()` is setup to wrangle the cluster assignments into the proper format for either `fit_model()` or `analyze_questioned_documents()`. Instead of adapting `get_clusterassignment()` to work on other types of data as well, I think it makes more sense to make a separate exported function that gets cluster assignments for docs in a user defined folder exported in a general format.