Halved the number of calls to `shortest.paths` in `AllUniquePaths`
## shortest.paths versus shortest_paths
The igraph functions `shortest.paths` and `shortest_paths` are not the same! The former is deprecated and has been replaced by `distances`. It retuns the distance(s) of the shortest path between pair(s) of nodes. The later returns the vertices and/or edges in shortest path(s).

## Halved the number of calls to `shortest.paths`
I updated the function to call `distances` instead of `shortest.paths`. And I changed the code so that `distances is called once per iteration of the while loop instead of twice.

## Other speed improvements?
`AllUniquePaths` returns a list of all paths between pairs of nodes. From the logic, I think these paths must be simple, i.e. they do not visit a node more than once. This would mean the paths also cannot be cycles because then the beginning and end node would be visted twice.

Ideas for speed improvement:
- `igraph::all_simple_paths` might produce the same result and run faster
- instead of searching the entire input graph for shortest paths, search connected subgraphs one-by-one. `igraph::components` returns the list of connected vertices. Then I think `igraph::induced_subgraph` could be used to get the subgraph for the connected vertices.

However, I am going to move on to more pressing problems.

## Unused argument graph
`AllUniquePaths` took two graphs as input, but only one is used so I dropped the unused graph as an argument.