Skip to next doc if error occurs in `batch_process_list()`
## Previous behavior
Before this commit, if `batch_process_list()` encountered an error on a document, the function would stop. The user could rerun `batch_process_list()` and the function would skip documents that had already been processed, but unless the user removed the problem document from the folder by hand, the function would try again to process the document that caused the error.

## Changes:
`batch_process_list()` performs the following steps:
1. Removes any documents that have already been processed and saved in output_dir from the images list (the docs to-be-processed list). If the images list is empty, `batch_process_list()` returns NULL.
2. Checks whether output_dir > problems.txt exists. This file records the error messages and document names of files that `batch_process_list()` was unable to process on previous runs. If the file exists, the file is loaded and the doc names are extracted into a vector.
3. If skip_docs_on_retry is TRUE, the problem docs are removed from the images list. Then if the images list is empty, `batch_process_list()` returns NULL. If skip_docs_on_retry is FALSE, `batch_process_list()` will leave the problem docs in the images list and try again to process them in the following step.
4. For each document in the images list, `batch_process_list()` tries to process the document inside a try-catch statement. If an error is encountered, the error message and doc name are written to the log file, and the function skips to the next document in the images list.
5. After every document in the images list has been either processed or flagged as a problem file in the log, the `batch_process_list()` returns NULL.

## Known Bug
I have been testing `batch_process_list()` on the 27 documents from CSAFE Handwriting Database writer 1. w0001_pLND_s01_r02.png, w0001_pLND_s01_r03.png, and a handful of other docs encounter a "vector memory has been exhausted" error. The first 2-3 times this error is encountered, `batch_process_list()` behaves as expected. It records the error and the doc name in the log file and skips to the next document on the list. However, on the 3rd or 4th time the vector memory error occurs, R aborts. In this case, the problem document is not recorded in the log file. Weirdly, when I rerun `batch_process_list()` it starts on the doc that caused R to abort, but instead of R aborting again, `batch_process_list()` records that there was an out of vector memory issue and skips to the next doc. It seems like R can only enounter the out of vector memory error a set number of times before it crashes. I tried using gc() but that does not fix the problem.