Stopped saving intermediate files in `make_clustering_templates()`
When I first tested and adapted Amy's clustering template code to put into handwriter, I saved several intermedate files so that processing wouldn't have to start from scratch if the function crashed. These files are large so they take a bit of time to save. `make_clustering_templates()` is stable, so we don't need the intermediate files for recovery. To save computational time, I updated the function to no longer save these intermediate files:
- template_proc_list.rds
- template_strata.rds
- template_images_list.rds

A few messages included time elapsed, but the majority of message did not. I removed time elapsed from the few messages that included it for consistency.