Changed to list by connected components and each component has its own list of skel_graph, graphdf, img_m, etc.
Previously, I had grouped features of connected components as follows:
- a master list 'comps' had a list of skel_graphs from all components
- 'comps' has another list of all graphdfs from all components
- 'comps' has yet another list of all paths from each component

Some of the lists contained in 'comps' were nested lists so it became difficult to keep track of which pieces belonged to the same component and which belonged to different components. In hopes that the change will make the component pieces easy to keep track of, I reoganized how the pieces are stored in the master list 'comps'. Now
- each item in 'comps' is a single connected component represented as a named list
- each connected component has a 'skel_graph', 'graph_df0', 'skel_graph0', 'img_m', 'adj0', 'indices', 'node_list', 'terminal_nodes', 'connected_nodes', etc.