Fixed errors from tests when running devtools::check()
## The Problem
Previously, the tests in test-slow-funcitons.R ran without errors with devtools::test() because it has access to all the template, model, and questioned doc examples in the examples folder. However, these tests failed when devtools::check() was run because devtools::check() runs off the built package which does not include the examples folder.

## The Solution
1. I removed some of the template, model, and questioned docs in the examples folder to make the file size less that 1 mb. I recreated example_cluster_template, example_model, and example_analysis using the paired-down docs.
2. I copied the template, model, and questioned docs to the fixtures folder in the tests folder where devtools::check() has access to them.
3. I updated the tests in test-slow-functions.R to use the template, model, and questioned docs in the test fixtures folder.

## Another Bug
Removing 4 of the 5 questioned documents to save space revealed a bug in calulcate_accuracy(). This function assumed the presence of a questioned document from each known writer. I updated calculate_accuracy() to work correctly even if some of the known writers did not have a corresponding questioned document.