Updated `plot_cluster_centers` to give user the option to only plot cluster centers
Older templates don't store the template graphs. New template might not store the template graphs in order to save space. If the template does not contain the template graphs, `plot_cluster_centers` no longer throws an error, even if the user tries to plot the template graphs along with the cluster centers.

Even if the template graphs are stored in the template, the user might want to only plot the cluster centers. This is now possible.