`get_clusters_batch` now skips docs without graphs
If `get_clusters_batch` encounters a processed document where doc$process$letterList is an empty list, it skips the doc and proceeds to the next.

A warning is given if the number of RDS files (excluding "all_clusters.rds" if it exists) in the output directory is not equal to the number of RDS files in the input directory.

If the function is running sequentially, a warning is given if a document does not contain any graphs.

I chose not to show a warning if the function is running sequentially because the foreach loop would need to be restructured which will take time. This could be done in the future.