Switched to styles.css
csafe.css is a huge file and I don't know which tags are used by the website. I switched to styles.css which is a bare bones css file. Then, I started updating the formatting in this file. In theory, every tag in this file should be will be used in the website.

## Formatting Changes

1. Increased the logo size
2. Changed the background color of the navbar to white
3. Aligned the navbar links vertically to the bottom
4. Changed the inactive navbar link color to light gray
5. Changed the active navbar link color to gray
6. Changed the hover navbar link color to light gray. This is in fact the same color as the inactive link color, but I'm leaving in the hover tag in case I want to change it later.