Fixed bug in website TOC
Previously, with styles.css the last link in the website table of contents was always highlighted, even if another link was active.

I switched back to csafe.css because it highlights the correct table of contents link. I added the lines of code to csafe.css that fix the bug where the inactive navbar links are white.