Adjusted plot format for `plot_graphs` to work with handwriterApp
The previous format of `plot_graphs` did not display properly when rendered in handwriterApp. The facet labels were huge and the graphs were miniscule. I tried removing the facet labels and added an argument for number of columns for the faceting. I will see if this works well in handwriterApp.