Added fixtures to make slow tests faster
The three tests in test-slow-functions.R were to slow to run on CRAN or CI. I added fixtures of the processed graphs and clusters. The tests run the functions in the temp dir, and the tests were slow because the graphs and clusters had to be processed each time. Now the tests copy the pre-processed graph and cluster files to the temp dir before running the functions to save time.